requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
psycopg2
python-dotenv
authlib
requests